+++ lib/jellyfish_azure/operation/azure_deprovision_operation.rb
@@ -1,69 +1,28 @@
 module JellyfishAzure
   module Operation
-    class AzureDeprovisionOperation
-      WAIT_TIMEOUT = 14_400
-      WAIT_DELAY = 15
-
-      attr_accessor :deploy_timeout, :deploy_delay
-
-      def initialize(cloud_client, provider, product, service)
-        @cloud_client = cloud_client
-        @provider = provider
-        @product = product
-        @service = service
-
-        @deploy_timeout = WAIT_TIMEOUT
-        @deploy_delay = WAIT_DELAY
-      end
-
-      def setup
-      end
-
-      def location
-      end
-
+    class AzureDeprovisionOperation < AzureOperation
       def execute
-        setup
-
+        log_note 'Deprovision started'
+        set_status :stopping,'Deprovision started'
         @cloud_client.resource_group.remove_resource_group  @service.resource_group_name
 
         check_status @service.resource_group_name
 
         set_status :terminated, 'Deprovision successful'
-
+        log_note 'Deprovision completed successfully'
       rescue WaitUtil::TimeoutError
-        handle_error ['The deprovisioning operation timed out.']
+        handle_errors ['The deprovisioning operation timed out.']
       rescue ValidationError => e
-        handle_error [e.to_s]
+        handle_errors [e.to_s]
       rescue AzureDeploymentErrors => e
-        handle_error e.errors.map(&:error_message)
+        handle_errors e.errors
       rescue MsRestAzure::AzureOperationError => e
         handle_azure_error e
       rescue => e
-        handle_error "Unexpected error: #{e.class}: #{e.message}"
+        handle_errors ["Unexpected error: #{e.class}: #{e.message}"]
       end
 
       private
-
-      def handle_error(messages)
-        set_status :terminated, messages[0]
-        messages.each do |msg|
-          entry = @service.logs.new
-          entry.update_attributes(log_level: 'error', message: msg) unless entry.nil?
-        end
-      end
-
-      def handle_azure_error(error)
-        msg = error.body.nil? ? error.message : error.body['error']['message']
-        set_status :terminated, msg
-        msg = error.body.nil? ? error.message : error.body['error']['message']
-        entry = @service.logs.new
-        entry.update_attributes(log_level: 'error', message: msg) unless entry.nil?
-      end
-
-      def set_status(status, message)
-        @service.update_attributes( status: status, status_msg: message )
-      end
 
       def check_status(resource_group_name)
         rg = @cloud_client.resource_group.get resource_group_name
@@ -71,16 +30,13 @@ module JellyfishAzure
           errors = AzureDeploymentError.new 'Unable to completely remove all resources in group'
           fail AzureDeploymentErrors, errors
         end
-      rescue AzureDeploymentErrors => e
-        handle_azure_error e
       rescue MsRestAzure::AzureOperationError => e
         # if resource group not found then delete was successful
+        # otherwise some error occurred.
         if  e.body['error']['code'] != 'ResourceGroupNotFound'
           msg = e.body.nil? ? e.message : e.body['error']['message']
-          handle_error [msg]
+          handle_errors [msg]
         end
-      rescue => e
-        handle_error ["Unexpected error: #{e.class}: #{e.message}"]
       end
     end
   end
