+++ lib/jellyfish_azure/operation/azure_provision_operation.rb
@@ -1,22 +1,7 @@
-
 module JellyfishAzure
   module Operation
-    class AzureProvisionOperation
+    class AzureProvisionOperation < AzureOperation
       DEPLOYMENT_NAME = 'Deployment'
-      WAIT_TIMEOUT = 14_400
-      WAIT_DELAY = 15
-
-      attr_accessor :deploy_timeout, :deploy_delay
-
-      def initialize(cloud_client, provider, product, service)
-        @cloud_client = cloud_client
-        @provider = provider
-        @product = product
-        @service = service
-
-        @deploy_timeout = WAIT_TIMEOUT
-        @deploy_delay = WAIT_DELAY
-      end
 
       def setup
       end
@@ -33,11 +18,11 @@ module JellyfishAzure
       def execute
         setup
 
+        set_status :provisioning, 'Provisioning service'
+
         @cloud_client.resource_group.create_resource_group @service.resource_group_name, location
 
         @cloud_client.deployment.create_deployment @service.resource_group_name, DEPLOYMENT_NAME, template_url, template_parameters
-
-        set_status :provisioning, 'Provisioning service'
 
         outputs = wait_for_deployment DEPLOYMENT_NAME
 
@@ -46,51 +31,18 @@ module JellyfishAzure
         set_status :available, 'Deployment successful'
 
       rescue WaitUtil::TimeoutError
-        handle_error ['The provisioning operation timed out.']
+        handle_errors ['The provisioning operation timed out.']
       rescue ValidationError => e
-        handle_error [e.to_s]
+        handle_errors [e.to_s]
       rescue AzureDeploymentErrors => e
-        handle_error e.errors
+        handle_errors e.errors
       rescue MsRestAzure::AzureOperationError => e
         handle_azure_error e
       rescue => e
-        handle_error ["Unexpected error: #{e.class}: #{e.message}"]
+        handle_errors ["Unexpected error: #{e.class}: #{e.message}"]
       end
 
       private
-
-      def handle_error(messages)
-        set_status :terminated, messages[0]
-        messages.each do |msg|
-          entry = @service.logs.new
-          entry.update_attributes(log_level: 'error', message: msg) unless entry.nil?
-        end
-      end
-
-      def handle_azure_error(error)
-        msg = error.body.nil? ? error.message : error.body['error']['message']
-        set_status :terminated, msg
-        msg = error.body.nil? ? error.message : error.body['error']['message']
-        entry = @service.logs.new
-        entry.update_attributes(log_level: 'error', message: msg) unless entry.nil?
-      end
-
-      def set_status(status, message)
-        @service.status = status
-        @service.status_msg = message
-        @service.save
-      end
-
-      def save_outputs(outputs)
-        outputs.each do |key, value|
-          service_output = get_output(key) || @service.service_outputs.new(name: key)
-          service_output.update_attributes(value: value[:value], value_type: :string) unless service_output.nil?
-        end
-      end
-
-      def get_output(name)
-        @service.service_outputs.where(name: name).first
-      end
 
       def wait_for_deployment(deployment_name)
         state = nil
